remove sky from requirements
