+++ requirements.txt
@@ -7,12 +7,13 @@ opentelemetry-instrumentation
 opentelemetry-instrumentation-fastapi
 opentelemetry-instrumentation-requests
 opentelemetry-sdk
+pandas
 pexpect
 pyarrow
 pyopenssl>=23.3.0
+ray
 rich
 sentry-sdk
-skypilot[docker]==0.4.1
 sshfs
 sshtunnel>=0.3.0
 typer
