+++ runhouse/resources/hardware/on_demand_cluster.py
@@ -5,10 +5,15 @@ import time
 from pathlib import Path
 from typing import Any, Dict
 
+# import sky here
 import rich.errors
-import sky
 import yaml
-from sky.backends import backend_utils, CloudVmRayBackend
+
+try:
+    import sky
+    from sky.backends import backend_utils, CloudVmRayBackend
+except ImportError:
+    pass
 
 from runhouse.globals import configs, rns_client
