+++ runhouse/servers/http/http_server.py
@@ -15,8 +15,6 @@ from fastapi import Body, FastAPI, HTTPException, Request
 from fastapi.encoders import jsonable_encoder
 from fastapi.responses import JSONResponse, StreamingResponse
 
-from sky.skylet.autostop_lib import set_last_active_time_to_now
-
 from runhouse.constants import (
     CLUSTER_CONFIG_PATH,
     DEFAULT_HTTP_PORT,
@@ -218,6 +216,8 @@ class HTTPServer:
 
     @staticmethod
     def register_activity():
+        from sky.skylet.autostop_lib import set_last_active_time_to_now
+
         set_last_active_time_to_now()
 
     @staticmethod
